Add gradle (and rename package)
