+++ .gitignore
@@ -21,3 +21,9 @@ text-ui-test/EXPECTED-UNIX.TXT
 /.jar/
 
 /data/
+
+# Ignore Gradle project-specific cache directory
+.gradle
+
+# Ignore Gradle build output directory
+build
