+++ src/main/java/Duke.java
@@ -1,3 +1,4 @@
+import duke.*;
 import duke.packages.*;
 
 import java.io.IOException;
