+++ src/main/java/duke/Deadline.java
@@ -1,4 +1,4 @@
-package duke.packages;
+package duke;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
