+++ src/main/java/duke/DukeException.java
@@ -1,4 +1,4 @@
-package duke.packages;
+package duke;
 
 public class DukeException extends Exception {
     String message;
