+++ src/main/java/duke/EmptyDescriptionException.java
@@ -1,4 +1,4 @@
-package duke.packages;
+package duke;
 
 public class EmptyDescriptionException extends DukeException {
     public EmptyDescriptionException(String message) {
