+++ src/main/java/duke/ToDo.java
@@ -1,4 +1,4 @@
-package duke.packages;
+package duke;
 
 public class ToDo extends Task {
     public ToDo(String description) {
